requirements.txt
streamlit
openai
langchain
python-dotenv
langchain-community
fpdf